database: remove redundant Qt::CaseInsensitive, change rowsAffected to qint64

+++ src/database/queryresult.h
@@ -16,7 +16,7 @@ struct QueryResult {
     bool isSelect = false;
     QStringList columns;
     QList<QueryRow> rows;
-    int rowsAffected = 0;
+    qint64 rowsAffected = 0;
 };
 
 #endif // QUERYRESULT_H
